Add quadax dependency, recover original files, remove jit for math_utils
