+++ pyproject.toml
@@ -105,6 +105,7 @@ dependencies = [ # Optional
   "dill",
   "jax",
   "numpy",
+  "quadax",
   "peppercorn",
   "sympy>=1.11"
 ]
